add back root layout and refactor to return <html>
